refactor: split repl into attach/repl subcommands
