+++ implementations/rust/ockam/ockam_command/src/zone/static/attach/long_about.txt
@@ -1,0 +1,1 @@
+This command can be used to access the Repl of an agent started locally.
